requirements.txt
uvicorn
fastapi
uvicorn[standard]
requests
msal
python-multipart
python-dotenv
itsdangerous
pydantic
starlette
openai